feat: comprehensive repo improvements
- Testing: Fix Firebase mocks, add Jest setup, update coverage threshold
- Code Quality: Standardize linting with --fix, enable TS strict mode
- Build & CI/CD: Add Turbo config, enhance GitHub Actions
- Debugging & Monitoring: Health checks, Pino logging, Sentry config
- Security & Performance: Vulnerability scanning, environment setup
- Documentation & Structure: Update README, add VS Code settings
- Development Workflow: Husky pre-commit hooks, branch protection

Resolves multiple issues: failing tests, lint errors, build failures, and improves overall DX.
