+++ apps/backend/package.json
@@ -7,7 +7,7 @@
   "license": "UNLICENSED",
   "main": "src/index.ts",
   "engines": {
-    "node": ">=18 <21"
+    "node": ">=18"
   },
   "scripts": {
     "build": "nest build",
