+++ apps/backend/src/api/__tests__/marketplace.test.ts
@@ -2,65 +2,6 @@ import request from 'supertest';
 import { db } from '../../config';
 import app from '../../functions/api';
 
-// Mock Firebase Admin SDK
-jest.mock('firebase-admin', () => ({
-  initializeApp: jest.fn(),
-  credential: {
-    applicationDefault: jest.fn(),
-  },
-  firestore: () => ({
-    collection: jest.fn(() => ({
-      get: jest.fn(() =>
-        Promise.resolve({
-          docs: [
-            {
-              id: 'creator1',
-              data: () => ({
-                portfolio: ['url1', 'url2'],
-                skills: ['skill1', 'skill2'],
-                rating: 4.5,
-              }),
-            },
-          ],
-        }),
-      ),
-      doc: jest.fn(() => ({
-        get: jest.fn(() =>
-          Promise.resolve({
-            exists: true,
-            id: 'creator1',
-            data: () => ({
-              portfolio: ['url1'],
-              skills: ['skill1'],
-              rating: 4.5,
-            }),
-          }),
-        ),
-        set: jest.fn(() => Promise.resolve()),
-        update: jest.fn(() => Promise.resolve()),
-      })),
-      add: jest.fn(() => Promise.resolve({ id: 'reviewId' })),
-    })),
-    terminate: jest.fn(() => Promise.resolve()),
-  }),
-  auth: () => ({}),
-}));
-
-// Mock multer
-jest.mock('multer', () => {
-  const multer = () => ({
-    single: () => (req: any, res: any, next: any) => {
-      req.file = {
-        buffer: Buffer.from('test'),
-        originalname: 'test.jpg',
-      };
-      next();
-    },
-  });
-  multer.memoryStorage = jest.fn();
-  return multer;
-});
-
 // Mock storage service
 jest.mock('../../services/storage', () => ({
   uploadBuffer: jest.fn().mockResolvedValue(undefined),
