+++ packages/ui/package.json
@@ -13,7 +13,7 @@
   "scripts": {
     "build": "tsc",
     "dev": "tsc --watch",
-    "lint": "eslint src --ext .ts,.tsx",
+    "lint": "eslint src --ext .ts,.tsx --fix",
     "clean": "rm -rf dist node_modules"
   },
   "peerDependencies": {
