+++ packages/ui/src/components/Button.tsx
@@ -1,14 +1,18 @@
 import React from 'react';
 import { cn } from '../utils';
 
-export interface ButtonProps extends React.ButtonHTMLAttributes<HTMLButtonElement> {
+export interface ButtonProps
+  extends React.ButtonHTMLAttributes<HTMLButtonElement> {
   variant?: 'primary' | 'secondary' | 'outline' | 'ghost';
   size?: 'sm' | 'md' | 'lg';
   children: React.ReactNode;
 }
 
 export const Button = React.forwardRef<HTMLButtonElement, ButtonProps>(
-  ({ className, variant = 'primary', size = 'md', children, ...props }, ref) => {
+  (
+    { className, variant = 'primary', size = 'md', children, ...props },
+    ref,
+  ) => {
     const variantStyles = {
       primary: 'bg-blue-600 text-white hover:bg-blue-700',
       secondary: 'bg-gray-600 text-white hover:bg-gray-700',
@@ -29,14 +33,14 @@ export const Button = React.forwardRef<HTMLButtonElement, ButtonProps>(
           'rounded-lg font-medium transition-colors',
           variantStyles[variant],
           sizeStyles[size],
-          className
+          className,
         )}
         {...props}
       >
         {children}
       </button>
     );
-  }
+  },
 );
 
 Button.displayName = 'Button';
