+++ packages/ui/src/components/Input.tsx
@@ -1,7 +1,8 @@
 import React from 'react';
 import { cn } from '../utils';
 
-export interface InputProps extends React.InputHTMLAttributes<HTMLInputElement> {
+export interface InputProps
+  extends React.InputHTMLAttributes<HTMLInputElement> {
   label?: string;
   error?: string;
 }
@@ -21,14 +22,14 @@ export const Input = React.forwardRef<HTMLInputElement, InputProps>(
             'w-full rounded-lg border border-gray-300 px-3 py-2 text-base',
             'focus:border-blue-500 focus:outline-none focus:ring-1 focus:ring-blue-500',
             error && 'border-red-500',
-            className
+            className,
           )}
           {...props}
         />
         {error && <p className="mt-1 text-sm text-red-500">{error}</p>}
       </div>
     );
-  }
+  },
 );
 
 Input.displayName = 'Input';
