+++ turbo.json
@@ -1,0 +1,24 @@
+{
+    "$schema": "https://turbo.build/schema.json",
+    "globalDependencies": [
+        "**/.env.*local"
+    ],
+    "tasks": {
+        "build": {
+            "dependsOn": [
+                "^build"
+            ],
+            "outputs": [
+                "dist/**",
+                ".next/**",
+                "!.next/cache/**"
+            ]
+        },
+        "lint": {},
+        "test": {},
+        "dev": {
+            "cache": false,
+            "persistent": true
+        }
+    }
+}
